requirements.txt
npmnumpy
pytest
ipython
